requirements.txt
discord.py
aiohttp
selenium